# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Nexora API
  version: 1.0.0
  description: 'Nexora is a multi-currency digital asset platform built on Stellar. It provides

    wallets, instant transfers, FX conversions, scheduled recurring payments,

    batch disbursements, fiat on/off-ramps, and a developer-friendly webhook system.


    All endpoints under `/v1` require a Bearer API key in the `Authorization` header.


    ## API key scopes


    An API key can be limited to specific resources and actions. Scopes have the

    form `<resource>:<action>`, where action is `read` or `write`:


    | Resource | Scopes | Covers |

    |---|---|---|

    | wallets | `wallets:read`, `wallets:write` | `/v1/wallets` (create, trustlines, contract wallet changes) |

    | transfers | `transfers:read`, `transfers:write` | `/v1/transfers`, `/v1/transactions`, `/v1/schedules`, `/v1/refunds` |

    | batches | `batches:read`, `batches:write` | `/v1/transfers/batch` |

    | webhooks | `webhooks:read`, `webhooks:write` | `/v1/webhooks` |

    | reports | `reports:read`, `reports:write` | `/v1/usage` (`reports:write` is reserved for report generation) |


    Other scopes: `keys:*`, `audit:read`, `fiat:*`, `fx:*`, `fees:read`,

    `beneficiaries:*`, `compliance:*`. `<resource>:*` grants both actions on one

    resource and `*` grants everything.


    On scoped resource groups, `GET`/`HEAD` need the `read` scope and every other

...

...

...

    path only; never the key itself).


...

...

...

...

...

...

    '
  contact:
...
servers:
...
security:
- BearerAuth: []
paths:
...
  /v1/wallets:
    post:
      operationId: createWallet
      summary: Create a wallet
      description: Generates a new Stellar keypair, encrypts the secret, and returns the wallet.
      tags:
      - Wallets
      responses:
        '201':
          description: Wallet created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletCreated'
              example:
                id: w_1a2b3c4d
                public_key: GBXCAFE7RJHBEVGT5ANBZ7YXDU7S3Y3VQWZ5J4YKJ2D6S7F8G9H0I
                created_at: '2026-01-15T10:00:00Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /v1/wallets/{id}:
    get:
      operationId: getWallet
      summary: Get wallet details
      description: Returns the wallet metadata including public key and creation timestamp.
      tags:
      - Wallets
      parameters:
      - $ref: '#/components/parameters/WalletID'
      responses:
        '200':
          description: Wallet details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wallet'
              example:
                id: w_1a2b3c4d
                public_key: GBXCAFE7RJHBEVGT5ANBZ7YXDU7S3Y3VQWZ5J4YKJ2D6S7F8G9H0I
                created_at: '2026-01-15T10:00:00Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/wallets/{id}/balances:
    get:
      operationId: getWalletBalances
      summary: Get wallet balances
      description: Returns all asset balances for the wallet from the Stellar network.
      tags:
      - Wallets
      parameters:
      - $ref: '#/components/parameters/WalletID'
      responses:
        '200':
          description: Wallet balances
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletBalances'
              example:
                wallet_id: w_1a2b3c4d
                balances:
                - asset_code: USDC
                  issuer: GA5ZSEJ6BHT2ZKGDABICHI62W25ZJUXUNFFYGLYTIEXP6CZMUWqmOvxM
                  balance: '1250.0000000'
                - asset_code: XLM
                  issuer: ''
                  balance: '42.5000000'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /v1/transfers:
    post:
      operationId: initiateTransfer
      summary: Initiate a transfer
      description: Moves an asset between two wallets, applying platform fees.
      tags:
      - Transfers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTransferRequest'
            example:
              from_wallet_id: w_1a2b3c4d
              to_wallet_id: w_5e6f7g8h
              asset: USDC
              amount: '100.0000000'
      parameters:
      - name: X-Idempotency-Key
        in: header
        required: false
        schema:
          type: string
        description: Optional idempotency key to prevent duplicate transfers on retries
      - name: Idempotency-Key
        in: header
        required: true
        description: 'Client-generated UUID v4 that makes this request safe to retry.

          The result is held for 24 hours: a repeat with the same body

          replays the original response, a repeat with a different body or

          a concurrent repeat is rejected with 409.

          '
        schema:
          type: string
          format: uuid
      responses:
        '202':
          description: Transfer accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferResponse'
              example:
                id: tx_abc123
                tx_hash: ''
                type: transfer
                status: pending
                from_wallet_id: w_1a2b3c4d
                to_wallet_id: w_5e6f7g8h
                asset: USDC
                amount: '100.0000000'
                fee_amount: '0.2500000'
                net_amount: '99.7500000'
                fee_bps: 25
                created_at: '2026-01-15T10:00:00Z'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
          description: 'Conflict: request in progress or idempotency key reused with different body'
    get:
      operationId: listTransfers
      summary: List transfers
      description: Returns all transactions for a given wallet.
      tags:
      - Transactions
      parameters:
      - name: wallet_id
        in: query
        required: false
        schema:
          type: string
          format: uuid
        description: Filter by wallet ID
      - name: external_reference
        in: query
        required: false
        schema:
          type: string
        description: Filter by external reference
      - name: tag
        in: query
        required: false
        schema:
          type: string
        description: Filter by tag
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
        description: Maximum number of results
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          default: 0
        description: Offset for pagination
      responses:
        '200':
          description: List of transactions
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactions:
                    type: array
                    items:
                      $ref: '#/components/schemas/TransferResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/transfers/{id}:
    get:
      operationId: getTransfer
      summary: Get transfer details
      description: Returns full details for a single transfer.
      tags:
      - Transfers
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: Transaction ID
      responses:
        '200':
          description: Transfer details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
      description: 'Pass a tenant API key (`sk_live_...` or `sk_test_...`) or JWT bearer token.

        In Swagger UI, enter your token in the Authorize dialog (the ''Bearer '' prefix is added automatically).

        '
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: 'Pass an API key as `Bearer <key>` in the `Authorization` header.

        '
    AdminAuth:
      type: http
      scheme: bearer
      bearerFormat: admin token
      description: Administrative bearer token.
  parameters:
    WalletID:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Wallet ID
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: BAD_REQUEST
              message: invalid request body
              status: 400
              request_id: req-example
    Unauthorized:
      description: Unauthorized — missing or invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: UNAUTHORIZED
              message: missing or invalid authorization header
              status: 401
              request_id: req-example
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: NOT_FOUND
              message: resource not found
              status: 404
              request_id: req-example
...
    Conflict:
      description: Conflict — a request with this Idempotency-Key is still in flight
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: REQUEST_IN_PROGRESS
              message: a request with this Idempotency-Key is still in progress
              status: 409
              request_id: req-example
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: INTERNAL_ERROR
              message: an unexpected error occurred
              status: 500
              request_id: req-example
  schemas:
...
    ErrorResponse:
      type: object
      required:
      - error
      properties:
        error:
          type: object
          required:
          - code
          - message
          - status
          - request_id
          properties:
            code:
              type: string
              enum:
              - BAD_REQUEST
              - ANCHOR_DEPOSIT_FAILED
              - ANCHOR_REGISTRATION_FAILED
              - ANCHOR_TRANSACTION_NOT_FOUND
              - ANCHOR_WITHDRAWAL_FAILED
              - API_KEY_NOT_FOUND
              - API_KEY_REVOKED
              - API_KEY_EXPIRED
              - BENEFICIARY_EXISTS
              - BENEFICIARY_NOT_ACTIVE
              - BENEFICIARY_NOT_ALLOWED
              - CONFIG_NOT_FOUND
              - CONFLICT
              - COOLING_OFF
              - CROSS_MODE_KEY_OPERATION
              - ENVIRONMENT_REQUIRED
              - IDEMPOTENCY_KEY_REQUIRED
              - IDEMPOTENCY_RECORD_PERSISTENCE_FAILED
              - IDEMPOTENCY_RESPONSE_TOO_LARGE
              - INSUFFICIENT_SCOPE
              - INVALID_ENVIRONMENT_MODE
              - INVALID_PREDICATE
              - INVALID_REQUEST
              - LAST_ORG_OWNER
              - METHOD_NOT_ALLOWED
              - MISSING_TRUSTLINE
              - NOT_IMPLEMENTED
              - PAYMENT_LINK_UNAVAILABLE
              - PREDICATE_NOT_SATISFIABLE
              - RATE_LIMITED
              - REFUND_CONFLICT
              - TEST_MODE_REQUIRED
              - TRANSFER_ID_INVALID
              - TRANSFER_ID_REQUIRED
              - TRANSFER_NOT_CANCELLABLE
              - UNSUPPORTED_FIAT_CURRENCY
              - WALLET_ID_REQUIRED
              - NOT_FOUND
              - UNAUTHORIZED
              - FORBIDDEN
              - INSUFFICIENT_SCOPE
              - INTERNAL_ERROR
              - INSUFFICIENT_SWEEPABLE_BALANCE
              - QUOTE_EXPIRED
              - QUOTE_ALREADY_USED
              - REVIEW_ALREADY_DECIDED
              - CLAIMABLE_BALANCE_NOT_PENDING
              - REQUEST_IN_PROGRESS
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error message
            status:
              type: integer
              description: HTTP status code; matches the response status
              minimum: 400
              maximum: 599
            request_id:
              type: string
              description: Effective X-Request-ID response header value
...
    Wallet:
      type: object
      properties:
        id:
          type: string
        public_key:
          type: string
          description: Stellar public key
        created_at:
          type: string
          format: date-time
    WalletCreated:
      type: object
      properties:
        id:
          type: string
        public_key:
          type: string
        created_at:
          type: string
          format: date-time
    WalletBalances:
      type: object
      properties:
        wallet_id:
          type: string
        balances:
          type: array
          items:
            $ref: '#/components/schemas/Balance'
    Balance:
      type: object
      properties:
        asset_code:
          type: string
          description: Asset code (e.g. USDC, XLM)
        issuer:
          type: string
          description: Issuer public key (empty for native XLM)
        balance:
          type: string
          description: Available balance as decimal string
...
    CreateTransferRequest:
      type: object
      required:
      - from_wallet_id
      - to_wallet_id
      - asset
      - amount
      properties:
        from_wallet_id:
          type: string
          format: uuid
        to_wallet_id:
          type: string
          format: uuid
        asset:
          type: string
          description: Asset code to transfer (e.g. USDC, XLM)
        amount:
          type: string
          description: Positive decimal amount
        external_reference:
          type: string
          nullable: true
          description: Optional client-specified reference for reconciliation
        tags:
          type: array
          items:
            type: string
          description: Optional tags for categorization and querying
    TransferResponse:
      type: object
      properties:
        id:
          type: string
        tx_hash:
          type: string
          description: Stellar transaction hash (empty until submitted)
        type:
          type: string
          enum:
          - transfer
          - conversion
          - funding
        status:
          type: string
          enum:
          - pending
          - submitted
          - confirmed
          - failed
          - reconciliation_failed
        from_wallet_id:
          type: string
        to_wallet_id:
          type: string
        asset:
          type: string
        amount:
          type: string
          description: Gross amount as decimal string
        fee_amount:
          type: string
          description: Fee deducted as decimal string
        net_amount:
          type: string
          description: Amount after fee as decimal string
        fee_bps:
          type: integer
          description: Fee in basis points
        external_reference:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        failure_reason:
          type: string
          description: Machine-readable reason code when the transfer failed during initiation
        failure_message:
          type: string
          description: Underlying initiation error for support diagnostics
        created_at:
          type: string
          format: date-time
...